requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
grpcio==1.54.2
grpcio-tools==1.54.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.23.0
Werkzeug==2.3.4
